Questions and answers populating on screen. Answers are buttons now, still unfunctional. Cleaned up characters in answers/question not printing properly (needs some work)
